requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
joblib==1.4.2
nltk==3.9.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
regex==2024.11.6
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
textblob==0.18.0.post0
tqdm==4.67.0
typing_extensions==4.12.2
uvicorn==0.32.0
python-decouple

